+++ doc/index.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html><head>
+  <title>Revisiting 2-3 Red-Black Trees with a Pedagogically Sound yet
+Efficient Deletion Algorithm: The Parity-Seeking Delete Algorithm</title>
+
+  
+  
+  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+
+  
+  <meta name="author" content="Kamaledin Ghiasi-Shirazi" />
+
+  
+  <meta name="author" content="Taraneh Ghandi" />
+
+  
+  <meta name="author" content="Ali Taghizadeh" />
+
+  
+  <meta name="author" content="Ali Rahimi-Baigi" />
+
+  
+  <meta name="keywords" content="red-black trees, 2-3 red-black trees, parity-seeking, delete algorithm, pedagogical, efficient" />
+
+  
+  <meta name="description" content="This page provides some summplementary material for the paper: 
+Ghiasi-Shirazi, Kamaledin, Taraneh Ghandi, Ali Taghizadeh, and Ali Rahimi-Baigi. " a="" pedagogically="" sound="" yet="" efficient="" deletion="" algorithm="" for="" red-black="" trees:="" the="" parity-seeking="" delete="" algorithm.="" arxiv="" preprint="" arxiv:2004.04344="" (2020)="" />
+
+  
+  <link rel="stylesheet" type="text/css" href="../publications.css" />
+
+</head><body>
+<p id="linkSeq">
+<a href="../../index.html">Home</a> <span> ? </span>
+<a href="../index.html">Publications</a> <span> ? </span>
+</p>
+
+<div id="header">
+<h1><span style="color: rgb(0, 0, 0);">Revisiting 2-3 Red-Black Trees
+with a Pedagogically Sound yet Efficient Deletion Algorithm: The
+Parity-Seeking Delete Algorithm</span></h1>
+</div>
+
+<div id="content">
+This page provides some supplementary material for the following
+publication:
+<p class="PaperRef">Ghiasi-Shirazi, Kamaledin, Taraneh Ghandi, Ali
+Taghizadeh, and Ali Rahimi-Baigi. A Pedagogically Sound yet Efficient
+Deletion algorithm for Red-Black Trees: The Parity-Seeking Delete
+Algorithm. <em>***</em>, Volume 00, Issue 00, pp 0000–0000, 2022.
+<a href="https://arxiv.org/abs/2004.04344">(link)
+</a></p>
+<h2 id="Software"><em>Implementation:</em></h2>
+<br />
+<a href="https://github.com/k-ghiasi/RedBlackTrees">https://github.com/k-ghiasi/RedBlackTrees</a><br />
+<br />
+<h2 id="Software"> <em>Detailed execution steps of different red-black
+tree algorithms: </em> </h2>
+<a href="clrs-insert-images.pdf">Insert of CLRS<br />
+</a><a href="clrs-insert-images.pdf">Delete of CLRS</a> <br />
+<br />
+<a href="llrb-insert-images.pdf">Insert of LLRB</a> <br />
+<a href="llrb-delete-images-part-I.pdf">Delete of LLRB (Part I)</a> <br />
+<a href="llrb-delete-images-part-II.pdf">Delete of LLRB (Part II)</a> <br />
+<br />
+<a href="2-3-psrbt-insert-images.pdf">Insert of 2-3 Parity-Seeking
+Red-Black Trees</a> <br />
+<a href="2-3-psrbt-delete-images.pdf">Delete of 2-3 Parity-Seeking
+Red-Black Trees</a> <br />
+<br />
+<a href="2-3-4-psrbt-insert-images.pdf">Insert of 2-3-4 Parity-Seeking
+Red-Black Trees (The same as CLRS)</a>
+<br />
+<a href="2-3-4-psrbt-delete-images.pdf">Delete of 2-3-4 Parity-Seeking
+Red-Black Trees</a>
+<br />
+</div>
+
+</body></html>
