+++ doc/publications.css
@@ -1,0 +1,156 @@
+/* Generated by KompoZer */
+html, body {
+  margin: 0;
+  padding: 0;
+}
+body {
+  background: #ffffff none repeat scroll 0% 50%;
+  font-family: TrebuchetMS,Arial,sans-serif;
+  font-style: normal;
+  font-variant: normal;
+  font-weight: normal;
+  font-size: 85%;
+  line-height: 1.3;
+  font-size-adjust: none;
+  font-stretch: normal;
+  text-align: left;
+  padding-bottom: 20px;
+}
+a {
+  color: #0066b3;
+  text-decoration: none;
+}
+h1 {
+  font-family: Arial,Arial,Sans-Serif;
+  font-style: normal;
+  font-variant: normal;
+  font-weight: bold;
+  font-size: 2.1em;
+  line-height: normal;
+  font-size-adjust: none;
+  font-stretch: normal;
+  color: #036da7;
+}
+h2 {
+  margin: 0;
+  padding: 0;
+  font-stretch: normal;
+  font-size-adjust: none;
+  line-height: normal;
+  font-size: 1.1em;
+  font-style: normal;
+  font-weight: bold;
+  font-family: Arial,Arial,Sans-Serif;
+  font-variant: normal;
+}
+#linkSeq {
+  margin-left: 2.2em;
+}
+#linkSeq a {
+  border:  none;
+  background: transparent url(../icons/rarr.gif) no-repeat scroll right center;
+  padding-right: 25px;
+}
+#linkSeq a:hover {
+  text-decoration: underline;
+}
+#linkSeq span {
+  position: absolute;
+  left: -999em;
+}
+div#header {
+  margin: 0 auto;
+  background: #bbd9ee none repeat scroll 0% 50%;
+  overflow: hidden;
+  width: 90%;
+}
+div#header h1, div#menu {
+  margin: 0 auto;
+  width: 1000px;
+}
+div#header h1 {
+  padding: 30px 0 20px;
+  color: #000066;
+  text-align: center;
+}
+ul#nav, ul#nav li {
+  margin: 0;
+  padding: 0;
+  list-style-type: none;
+}
+ul#nav {
+  float: right;
+  font-size: 80%;
+}
+ul#nav li {
+  float: left;
+  margin-left: 3px;
+  text-align: center;
+}
+ul#nav a {
+  border-top: 1px solid #ffffff;
+  border-left: 1px solid #ffffff;
+  border-right: 1px solid #ffffff;
+  padding: 5px 0;
+  background: #e7f1f8 none repeat scroll 0% 50%;
+  float: left;
+  width: 95px;
+  text-decoration: none;
+  color: #666666;
+}
+ul#nav a:hover {
+  background: #ffffff none repeat scroll 0% 50%;
+  color: #000000;
+}
+ul#nav li.activelink a, ul#nav li.activelink a:hover {
+  background: #ffffff none repeat scroll 0% 50%;
+  color: #000033;
+}
+#content {
+  margin: 0 auto;
+  padding: 15px;
+  background: #eeeeee none repeat scroll 0% 50%;
+  font-size: 100%;
+  font-style: normal;
+  font-size-adjust: none;
+  line-height: 1.3;
+  width: 95%;
+  font-weight: normal;
+  font-family: TrebuchetMS,Arial,sans-serif;
+  font-variant: normal;
+  text-align: justify;
+  font-stretch: normal;
+}
+#content h2 {
+  color: #3333ff;
+}
+.PaperRef {
+  font-style: normal;
+  font-weight: bold;
+  color: #006600;
+}
+.datasetTable {
+  border: 1px solid #ff99ff;
+  margin: auto;
+  border-collapse: collapse;
+  text-align: left;
+}
+.datasetTable th, td {
+  border: 1px solid #ff99ff;
+  border-collapse: collapse;
+  text-align: center;
+}
+.datasetTable caption {
+  margin: 0 auto 4pt;
+  font-weight: bold;
+  color: #336666;
+}
+#content h3 {
+  color: #3366ff;
+  font-style: italic;
+  margin-left: 8px;
+}
+dt {
+  font-weight: bold;
+  padding-left: 12px;
+}
